refactor(web): split FilesPage into focused files/ components
Bring FilesPage under the react-doctor size limit by extracting toolbar, search, views, and delete dialog the same way AutomationsPage does.
